Adding Day 2/3
Today I finished Days 2 and 3. In day 2, I covered variables, checked data types, use len() function, calculated the area of a circle and used took user input. In day 3, i calculated more areas and used the 'and' and 'in' operator to check for words or characters in a string. I also worked more with types and how int conversion rounds down. Finally, I ended day 3 with a script that finds the pay of a person and the amount of seconds a person has lived for given the years they lived