Day 13: List Comprehension
Today I learned list comprehension. To my understanding, it allows your code to be more legible and concise by adding the for loop into the list so that there aren't so many lines of code. I also learned about the lambda function. It basically is a mini function that can take parameters. This makes it useful for quick and simple operations without having to define a function.