Day 18: Regular Expressions
Today I learned about the re module. Previously, I had no experience, but today I learned the useful aspects of the re module. Today I used the refunction for find all words in a list through the \b\w+\b re pattern.  I also was able to use the re to check whether or not an expression is a valid python name. I was able to achieve this by creating a function that only took upper case and lower case numbers and underscores. Following that, there could be numbers letters or underscores. I made sure to include the * at the end to represent 0 or more  because more characters than 1 is considered optional. An example of this is 'a' and 'abc.'  I also learned how to use sub in order to clean a string that had various symbols that needed replacing.