Day 22: Web Scraping
Today was one of the most interesting days during this course. Here I was able to go to websites and directly take data from websites. Today was also more difficult because I had no knowledge about HTML tags. However, I was able to figure it out and create web scraping programs. The first exercise is filled with comments as I wanted to understand each code of line. In the following practices, I chose different websites and I added them to the json in the jsons folder.