Day 23: Virtual Environment
Today I covered the topic of virtual environments. Essentially, virtual environments are environments that can be made within projects to ensure that multiple projects do not have issues relating to different packages. In addition to that, I used flask to create website that said 'Hello, Flask!' I wanted to take it a step further and added a book scraper. Essentially, you would open the website and type in a name of a book. Then that would go to another website enter that into the the web browser and would fetch all of the names of the books and its authors.