Day 24: Statistics
Today, I covered the statistics aspect of python. I had to download Anaconda and I used Jupyter. In addition, I practiced using numpy. The numpy practice was relatively basic, but the more interesting part was the fact that I was able to create graphs based off of randomly generated data.