Day 27/28: Python With MongoDB
I combined both of these days so that I could have two days instead of one day to build an API. Starting with day 27, I learned how to use and add to a database as well as remove/find from a database. For day 28, it covered the basics of an API. An API is an application program interface that utilizes GET, PUT, POST, and delete to manage data to allow different systems to communicate using HTTP methods.